Quick change to Player movement (see description)
Made it so that if you're holding down one directional and you press the
opposite directional, the player will change direction.

TODO: Make it so that if you are holding down a key, press the opposite
key, then release it, the player will move one direction, then move in
the opposite direction, then move back to the original direction.